requirements.txt
Flask==2.2.3
requests==2.28.1
opencensus==0.11.0
opencensus-ext-azure==1.0.8  
azure-monitor-query==1.0.0
azure-identity==1.7.1
gunicorn==20.1.0
psutil==5.9.3  
